Ajeitando diversos bugs, comentando lugares onde pode ter mais

+++ Camera/Camera.cpp
@@ -9,7 +9,7 @@ Camera::Camera(Vec3 position, double fov, Vec3 target, double near, Vec3 up, int
     this->axisX = axisZ.cross(up, axisZ).getUnitVector();
     this->axisY = axisZ.cross(axisZ, axisX).getUnitVector();
 
-    this->aspect = width/height;
+    this->aspect = (float) width/ (float )height;
     this->near = near;
     this->fov = fov*M_PI/180;
     this->position = position;
@@ -19,11 +19,13 @@ Ray Camera::getRay(double x, double y, int width, int height) const {
     double halfHeight = this->near * tanf(this->fov/2);
     double halfWidth = aspect * halfHeight;
 
+    // Erro pode estar aqui - Temos que mudar a base
     double pointX = halfHeight * (2.0 * ((x/width) - 0.5 ));
     double pointY = -halfWidth  * (2.0 * ((y/height) - 0.5));
     double pointZ = -this->near;
 
     Vec3 point(pointX, pointY, pointZ);
+
 
     return Ray(position, point-position);
 }
